% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeType.R
\name{attributeType}
\alias{attributeType}
\title{Provide a rich description of the variable's data type}
\usage{
attributeType(x)
}
\arguments{
\item{x}{A value of which the type has to be determined}
}
\value{
A list of detailed description of \code{x}'s data type. The list always contains the
  \code{type} element, which contains:
  \itemize{
    \item \code{mode(x)} if \code{x} is an atomic data value (such as character, numeric etc.);
    \item \code{class(x)} if \code{x} is an S3 object.
  }
  Additionally, the returned list may contain other elements such as \code{levels} for factors.
}
\description{
Function is used in DAO objects within the \code{getAttributes} method - see e.g.
\code{\link{sqlDao}}.
}
