% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crudTable.R
\name{crudTableUI}
\alias{crudTableUI}
\title{Creates a data table view that allows to add new data, modify existent records or delete the records.}
\usage{
crudTableUI(
  id,
  newButtonLabel = "New Record",
  newButtonIcon = icon("plus"),
  newButtonClass = "btn-primary",
  newButtonWidth = NULL
)
}
\arguments{
\item{id}{The ID of the widget}

\item{newButtonLabel}{Label of the button for adding of new records. If 'NULL', the button is not shown.}

\item{newButtonIcon}{Icon of the button for adding of new records}

\item{newButtonClass}{Class of the button for adding of new records}

\item{newButtonWidth}{The width of the button, e.g. '400px' or '100%'}
}
\value{
An editable data table widget
}
\description{
Creates a data table view that allows to add new data, modify existent records or delete the records.
}
