% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editDialog.R
\name{editDialog}
\alias{editDialog}
\title{Edit dialog for the create and update operations of the \code{\link{crudTable}}}
\usage{
editDialog(
  id,
  ...,
  newTitle = "New",
  editTitle = "Edit",
  submitLabel = "Submit",
  cancelLabel = "Cancel"
)
}
\arguments{
\item{id}{Shiny module identifier of the edit dialog as set by the \code{\link{crudTable}}. (The id is based on
the \code{\link{crudTable}}'s namespace and ends either with \code{"newForm"} or \code{"editForm"} accordingly
to whether the dialog is used for new data input or existing data update.)}

\item{...}{The definition of the edit dialog, i.e. the shiny input elements whose IDs must be named after
the attribute names, as expected by the DAO (see \code{\link{dataFrameDao}} or \code{\link{sqlDao}} for
more information on Data Access Objects). IDs must also be namespaced, as need by the UI modules, see
the example below.}

\item{newTitle}{A dialog title used for new data input}

\item{editTitle}{A dialog title used for existing data update}

\item{submitLabel}{Label of the submit button}

\item{cancelLabel}{Label of the cancel button}
}
\description{
Edit dialog for the create and update operations of the \code{\link{crudTable}}
}
\examples{
\dontrun{
# A typical use of editDialog - create a form UI for the crudTable:
formUI <- function(id) {
   # create namespace - note the use of ns() below in *Input calls
   ns <- NS(id)
   editDialog(id,
              textInput(ns('Plant'), 'Plant'),
              selectInput(ns('Type'), 'Type', choices = c('Quebec', 'Mississippi')),
              selectInput(ns('Treatment'), 'Treatment', choices = c('nonchilled', 'chilled')),
              numericInput(ns('conc'), 'Ambient CO2 concentration [ml/L]', value = 100, min = 50, max = 1000),
              numericInput(ns('uptake'), 'CO2 uptake rates [umol/m2 sec]', value = 0, min = 0, max = 100),
   )
}
}
}
\seealso{
crudTable
}
