% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filledValidator.R
\name{filledValidator}
\alias{filledValidator}
\title{Checks if the inputs are not empty.}
\usage{
filledValidator(inputIds, errorMessages = "Must not be empty.")
}
\arguments{
\item{inputIds}{IDs of shiny inputs that are to be tested for emptiness}

\item{errorMessages}{a vector of error messages corresponding to the input IDs. This character
vector is recycled as needed to match the length of \code{inputIds}}
}
\value{
A list of instances of the S3 class \code{validator}. The size of the resulting list
    equals to the number of input IDs in 'inputIds'.
}
\description{
This validator tests if the inputs given by 'inputIds' are not empty, i.e., non-null,
non-NA and different from an empty string.
}
\seealso{
validator
}
