% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDao.R
\name{is.dao}
\alias{is.dao}
\title{Test whether the given object is a correct 'dao' class}
\usage{
is.dao(x)
}
\arguments{
\item{x}{An object to be tested}
}
\value{
\code{TRUE} if 'x' is a valid DAO object.
}
\description{
Test whether the given object is a correct 'dao' class
}
\seealso{
dataFrameDao, sqlDao
}
