% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTypecast.R
\name{is.typecast}
\alias{is.typecast}
\title{Test if the argument is a valid typecast object}
\usage{
is.typecast(x)
}
\arguments{
\item{x}{An object to be tested}
}
\value{
\code{TRUE} if \code{x} is a valid instance of the \code{typecast} class.
}
\description{
Test if the argument is a valid typecast object
}
\seealso{
\code{\link{typecast}}
}
