% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidator.R
\name{is.validator}
\alias{is.validator}
\title{Checks whether the given argument is a valid variable object}
\usage{
is.validator(v)
}
\arguments{
\item{v}{An object to be tested}
}
\value{
\code{TRUE} if 'v' is a valid \code{\link{validator}}
}
\description{
This function tests if \code{v} is a list of objects that are instances of the S3 class
\code{\link{validator}} and whether each instance contains an 'errorMessage' character
scalar and a function 'f' as an element.
}
\seealso{
\code{\link{validator}}, \code{\link{filledValidator}}
}
