% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlDao.R
\name{sqlDao}
\alias{sqlDao}
\title{A Data Access Object (DAO) that uses the DBI interface to access the data.}
\usage{
sqlDao(con, table, typecast = list())
}
\arguments{
\item{con}{A DBI connection}

\item{table}{A character string of the name of the table to be accessed}
}
\value{
A DAO object, i.e. a list of functions for CRUD operations on the DBI table as neeeded by
  the \code{\link{crudTable}} module
}
\description{
DAO is a list that provides basic backend CRUD functionality to the \code{\link{crudTable}}. This
DAO uses DBI to store the data. The DBI table accessed with this DAO object must not contain the
\code{'id'} attribute, as it is internally created from the DBI's \code{'rowid'} attribute.
}
\details{
See \code{\link{dataFrameDao}} for more details on Data Access Objects.
}
\examples{
\dontrun{
library(DBI)
library(RSQLite)

# Create an in-memory database
con <- dbConnect(RSQLite::SQLite(), ":memory:")

# Create CO2 data table from the CO2 data frame
dbWriteTable(con, 'CO2', as.data.frame(CO2[1:5, ]))

# Create Data Access Object
dao <- sqlDao(con,
              table = 'CO2')
}
}
\seealso{
\code{\link{dataFrameDao}}, \code{\link{is.dao}}
}
