% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecast.R
\name{typecast}
\alias{typecast}
\title{Create a custom typecast object}
\usage{
typecast(fromInternal, toInternal)
}
\arguments{
\item{fromInternal}{a function for conversion of data from the internal to the output data format}

\item{toInternal}{a function for conversion of data from the output to the internal data format}
}
\value{
An object of class \code{'typecast'}, which is a list of two single-argument functions.
}
\description{
This function creates a typecast object, i.e. an object that provides functions for conversion
from and to the internal data storage format. It is used by \code{\link{sqlDao}} to transparently
convert between the internal and output data format.
}
\examples{
# Create a typecast object for data stored internally as character, but output as numeric values
tc <- typecast(as.numeric, as.character)
}
\seealso{
\code{\link{sqlDao}}, \code{\link{typecastDateToCharacter}},
  \code{\link{typecastDateToNumeric}}
}
