% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecastDateToCharacter.R
\name{typecastDateToCharacter}
\alias{typecastDateToCharacter}
\title{Create a typecast object that converts between the output \code{\link{Date}} and internal
\code{character} data type}
\usage{
typecastDateToCharacter(format = "\%Y-\%m-\%d")
}
\value{
An instance of the \code{'typecast'} class
}
\description{
Create a typecast object that converts between the output \code{\link{Date}} and internal
\code{character} data type
}
\seealso{
\code{\link{typecast}}, \code{\link{typecastDateToNumeric}}
}
