% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecastDateToNumeric.R
\name{typecastDateToNumeric}
\alias{typecastDateToNumeric}
\title{Create a typecast object that converts between the output \code{\link{Date}} and internal
\code{numeric} data type}
\usage{
typecastDateToNumeric()
}
\value{
An instance of the \code{'typecast'} class
}
\description{
Create a typecast object that converts between the output \code{\link{Date}} and internal
\code{numeric} data type
}
\seealso{
\code{\link{typecast}}, \code{\link{typecastDateToCharacter}}
}
