% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validator}
\alias{validator}
\title{Defines a validation function and an error message for an edit dialog form input}
\usage{
validator(inputIds, errorMessages, f)
}
\arguments{
\item{inputIds}{a vector of input IDs for the validation function \code{f} to be bound on}

\item{errorMessages}{a vector of error messages corresponding to the input IDs. This character
vector is recycled as needed to match the length of \code{inputIds}}

\item{f}{a validation function that expects a single input argument and returns \code{TRUE} if
the given input is valid. Note that this function must handle correctly the \code{NULL}
value as well as the \code{NA} value.}
}
\value{
A list of instances of the S3 class \code{validator}. The size of the resulting list
    equals to the number of input IDs in 'inputIds'.
}
\description{
Validators are typically passed to the \code{\link{formServerFactory}} function that
creates a server-part handler of the edit dialog.
}
\examples{
    validator('attr1',
              function(v) { !is.null(v) && !is.na(v) && v \%\% 2 != 0 })
}
\seealso{
\code{\link{filledValidator}}, \code{\link{formServerFactory}}
}
