% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editDialogServer.R
\name{editDialogServer}
\alias{editDialogServer}
\title{Server part of the edit dialog module for the create and update operations of
the \code{\link{crudTable}}}
\usage{
editDialogServer(attributes, validators = list())
}
\arguments{
\item{attributes}{A character vector of attribute names that correspond to IDs of shiny
inputs (as created in the \code{\link{editDialog}}) and as expected by the underlying
DAO}
}
\value{
A function that is used by shiny to handle the inputs of the edit dialog.
}
\description{
The result of the call is a function that handles the edit dialog on the server part of the
shiny application. Its purpose is to be passed as the 'formServer' argument of the
\code{\link{crudTable}} module. It handles the load and storing of record values that
are persisted via the \code{\link{crudTable}}'s DAO.
}
